oldapp/requirements.txt
chainlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
langchain-pinecone
langchain-google-genai
python-dotenv
pymongo
asyncpg
pyaudio
python-multipart
bcrypt